+++ angular_compiler/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   `CompilerFlags` no longer parses and supports the `'debug'` option and
+    `genDebugInfo` is always `false`, and is deprecated pending removal in a
+    future version.
+
 ## 0.4.0-alpha+14
 
 *   Maintenance release.